requirements.txt
python-telegram-bot==13.15
openai==0.27.0
firebase-admin==6.0.0
